Add files via upload
The objective of this project is to create a playful and engaging web application that generates random jokes on button click, while providing an interactive and fun user experience. A unique feature of this project is the addition of animated laughing emojis that fall from the top of the screen when a joke is displayed.